requirements.txt
flask==3.0.3
requests==2.31.0
openai==1.30.5
streamlit==1.45.1
transformers==4.41.2
torch==2.6.0
pillow==10.4.0
speechrecognition==3.10.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
python-dotenv==1.0.1
flask-migrate==4.0.7
gtts==2.5.4
flask-jwt-extended==4.6.0
numpy==1.26.4
python-magic==0.4.27
streamlit-webrtc==0.49.0
av==12.0.0
soundfile==0.12.1
werkzeug==3.0.3
pytesseract==0.3.10
pdf2image==1.17.0
langdetect==1.0.9
deep-translator==1.11.4
cachetools==5.3.3
packaging==24.0
click==8.1.7
streamlit-realtime-audio-recorder==0.1.2
pydub==0.25.1